requirements.txt
absl-py==2.3.1
certifi==2025.7.9
charset-normalizer==3.4.2
colorama==0.4.6
dnspython==2.7.0
filelock==3.18.0
fsspec==2025.7.0
grpcio==1.74.0
idna==3.10
Jinja2==3.1.6
Markdown==3.8.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
packaging==25.0
pillow==11.3.0
protobuf==6.31.1
pymongo==4.13.2
python-dotenv==1.1.1
requests==2.32.4
setuptools==80.9.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
torch==2.7.1
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
